ai-services/gfpgan/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
gfpgan==1.3.8
basicsr==1.4.2
facexlib==0.3.0
Pillow==11.1.0
numpy==2.2.2
opencv-python-headless==4.10.0.84
torch==2.5.1
torchvision==0.20.1
realesrgan==0.3.0
